Kvichak River SR Data

Spawner-recruit data for sockeye salmon from the Kvichak River.

Usage

data(KvichakSockeye)

Format

Object of class "data.frame"

brood_year Brood year.

spawners Spawners (count) in thousands

recruits Recruits (count) in thousands

pdo_summer_t2 Pacific Decadal Oscillation (PDO) from Apr-Sep of brood year $t+2$

pdo_winter_ts Pacific Decadal Oscillation (PDO) from Oct (brood year $t+2$) to Mar (brood year $t+3$)

Source

RAM Legacy Stock Assessment Database

Details

Spawner-recruit data for sockeye salmon (Oncorhynchus nerka) from the Kvichak River in SW Alaska that span the years 1952-1989. The data come from a large public database begun by Ransom Myers many years ago. The database is now maintained as the RAM Legacy Stock Assessment Database.

References

RAM Legacy Stock Assessment Database. 2018. Version 4.44-assessment-only. Released 2018-12-22. Retrieved from DOI:10.5281/zenodo.2542919.

Ricard, D., Minto, C., Jensen, O.P. and Baum, J.K. (2012) Evaluating the knowledge base and status of commercially exploited marine species with the RAM Legacy Stock Assessment Database. Fish and Fisheries 13 (4) 380-398. DOI: 10.1111/j.1467-2979.2011.00435.x

Spawner escapement estimates are originally from: 1997-2017 Appendix Table A12.

See also

Examples

data(KvichakSockeye)