Mauna Loa C02 measurements

Monthly Snow Water Equivalent (SWE) at Stations in Washington State

Usage

data(MLCO2)

Format

Objects of class "data.frame". Columns are year, month, ppm (part per million)

Source

NOAA ESRL Trends in Atmospheic Carbon Dioxide

Details

These are monthly average C02 values recorded on Mauna Loa in Hawaii since 1958. Data are provided by NOAA Earth System Rearch Laboratory, Global Monitoring Division.

Data from March 1958 through April 1974 have been obtained by C. David Keeling of the Scripps Institution of Oceanography (SIO) and were obtained from the Scripps website (scrippsco2.ucsd.edu). The ppm column contains the monthly mean CO2 mole fraction determined from daily averages. The mole fraction of CO2, expressed as parts per million (ppm) is the number of molecules of CO2 in every one million molecules of dried air (water vapor removed). If there are missing days, the value is the interpolated value. See www.esrl.noaa.gov/gmd/ccgg/trends/ for additional details.

References

Dr. Pieter Tans, NOAA/ESRL (www.esrl.noaa.gov/gmd/ccgg/trends/) and Dr. Ralph Keeling, Scripps Institution of Oceanography (scrippsco2.ucsd.edu/).

Examples

# We downloaded this data and created the `MLCO2` dataframe with the following code:
if (FALSE) { # \dontrun{
library(RCurl)
## get CO2 data from Mauna Loa observatory
ww1 <- "ftp://aftp.cmdl.noaa.gov/products/"
ww2 <- "trends/co2/co2_mm_mlo.txt"
CO2fulltext <- getURL(paste0(ww1,ww2))
MLCO2 <- read.table(text=CO2fulltext)[,c(1,2,5)]
## assign better column names
colnames(MLCO2) <- c("year","month","ppm")
save(MLCO2, file="MLCO2.RData")
} # }

data(MLCO2)