Lake Barco Aquatic Data

Lake Barco data for the EFI 2021 Forecasting Challenge

Usage

data(neon_barc)

Format

neon_barc is a tibble with columns are date, siteID, oxygen, temperature, oxygen_sd, temperature_sd, depth_oxygen, depth_temperature, neon_product_ids, year, and cal_day.

Source

https://github.com/eco4cast/neon4cast-aquatics

Details

The data for the aquatics challenge comes from a NEON site at Lake Barco (Florida). More about the data and challenge is here and the Github repository for getting all the necessary data is eco4cast.

The neon_barc data set was created with the aquatics-targets.csv.gz file produced in the neon4cast-aquatics GitHub repository and saved in the inst folder in the atsalibrary package. From that file, neon_barc is created with

library(tidyverse)
library(lubridate)
# This taken from code on NEON aquatics challenge
targets <- readr::read_csv("aquatics-targets.csv.gz", guess_max = 10000)
site_data_var <- targets %>%
  filter(siteID == "BARC")
# This is key here - I added the forecast horizon on the end of the data for the forecast period
full_time <- tibble(time = seq(min(site_data_var$time), max(site_data_var$time), by = "1 day"))
# Join the full time with the site_data_var so there aren't gaps in the time column
site_data_var <- left_join(full_time, site_data_var) %>%
  dplyr::rename(date = time)
site_data_var$year <- year(site_data_var$date)
site_data_var$cal_day <- yday(site_data_var$date)
neon_barc <- site_data_var

References

Ecological Forecasting Initiative https://ecoforecast.org/

Neon4Cast Aquatics GitHub repository https://github.com/eco4cast/neon4cast-aquatics

Examples

data(neon_barc)