Washington State SNOTEL measurements

Monthly Snow Water Equivalent (SWE) at Stations in Washington State

Usage

data(snotel)

snotel

snotelmeta

Format

snotel: object of class "data.frame". Columns are Station, Station.Id, Year, Month, SWE, and Date

snotelmeta: Station_Name, Station.Id, State.Code, Network.Code, Network.Name, Elevation, Latitude, Longitude"

An object of class data.frame with 27720 rows and 6 columns.

An object of class data.frame with 70 rows and 8 columns.

Source

USDA WA Snow Survey Program

Details

These data are the snow water equivalent percent of normal. This represents the snow water equivalent (SWE) compared to the average value for that site on the same day. The data were downloaded from the USDA Natural Resources Conservation Service website for the Washington Snow Survey Program. The snotel object is the SWE data and the snotelmeta is the station metadata (lat/lon and elevation).

Examples

data(snotel)